requirements.txt
psycopg2-binary==2.8.4
python-intercom==3.1.0
requests==2.22.0
SQLAlchemy==1.3.11
